web reader: mobile parity — scroll only, no hotkeys
- drop paginated render path, usePagination, useFullscreen, page-nav arrows
- single scroll path on all breakpoints, 680px centered column
- keyboard: Esc closes drawers, Cmd/Ctrl+F opens search
- auto-hide bars port mobile 48/6px scroll-dir threshold + 3s timer
- remove fullscreen button + shortcuts modal
- e2e: drop page-turn, use scroll to drive progress
