+++ apps/web/e2e/tests/reader-mobile.spec.ts
@@ -4,18 +4,17 @@ import { getTestData } from '../fixtures/test-data'
 import { waitForReaderLoad } from '../helpers/reader'
 
 test.describe('Mobile Reader', () => {
-  test('reader uses scroll mode on mobile', async ({ authedPage: page }) => {
+  test('reader uses scroll mode on all breakpoints', async ({ authedPage: page }) => {
     const { enBook } = getTestData()
     await page.goto(`/en/books/${enBook.slug}/${enBook.firstChapterSlug}`)
     await waitForReaderLoad(page)
 
-    // Mobile reader should use scroll mode
-    const scrollMode = page.locator('.reader-page--scroll-mode')
-    await expect(scrollMode).toBeVisible()
+    const scrollContainer = page.locator('.scroll-reader')
+    await expect(scrollContainer).toBeVisible()
 
-    // No pagination buttons in scroll mode
-    const pageNav = page.locator('.reader-page-nav')
-    await expect(pageNav).not.toBeVisible()
+    // Legacy pagination artefacts must be gone
+    await expect(page.locator('.reader-page-nav')).toHaveCount(0)
+    await expect(page.locator('.reader-page--scroll-mode')).toHaveCount(0)
   })
 
   test('auto-save on scroll (mobile)', async ({ authedPage: page }) => {
