--- apps/web/src/hooks/useFullscreen.ts
@@ -1,28 +1,0 @@
-import { useState, useEffect, useCallback } from 'react'
-
-export function useFullscreen() {
-  const [isFullscreen, setIsFullscreen] = useState(false)
-
-  useEffect(() => {
-    const handleChange = () => {
-      setIsFullscreen(!!document.fullscreenElement)
-    }
-
-    document.addEventListener('fullscreenchange', handleChange)
-    return () => document.removeEventListener('fullscreenchange', handleChange)
-  }, [])
-
-  const toggle = useCallback(() => {
-    if (!document.fullscreenElement) {
-      document.documentElement.requestFullscreen().catch(() => {
-        // Fullscreen not supported or denied
-      })
-    } else {
-      document.exitFullscreen().catch(() => {
-        // Already exited
-      })
-    }
-  }, [])
-
-  return { isFullscreen, toggle }
-}
